new video dataset format support and ds example

+++ torch_sensor_lib/sim.py
@@ -21,7 +21,7 @@ class FiberSimulator():
     def __init__(self, config, device='cpu'):
 
         self.device = device
-        self.dtype = np.float32
+        self.dtype = np.float16
         self.derivative_kernel = torch.from_numpy(
             np.expand_dims([1, -2, 1], (0, 1, 3)).astype(self.dtype)).to(device)
         self.config = config
